Proposed example format for epic details
This is a first attempt at specifying what a useful yaml descriptor
might look like for an OSIC Epic card, including task breakdowns.

It is modeled on the current status of the "Centralize configuration
defaults" epic, tagged CFGDFT on the Newton cycle's Trello board
titled: OSIC - Ceilometer/Horizon/FleetMgt.

An attempt has been made to make the yaml prodecurally generatable
and yet not include any unnecesarry elements as much as possible and
to structure the yaml in a hierarchial format that reflects the
actual organization of data within Trello's UI to allow for easy
correlation and facilitate a consistent mental model.

Feedback is desired.